+++ system/src/Grav/Common/Grav.php
@@ -48,6 +48,7 @@ use Grav\Common\Service\TaskServiceProvider;
 use Grav\Common\Twig\Twig;
 use Grav\Framework\DI\Container;
 use Grav\Framework\Psr7\Response;
+use Grav\Framework\RequestHandler\Middlewares\MultipartRequestSupport;
 use Grav\Framework\RequestHandler\RequestHandler;
 use Grav\Framework\Route\Route;
 use Grav\Framework\Session\Messages;
@@ -117,6 +118,7 @@ class Grav extends Container
      * @var array All middleware processors that are processed in $this->process()
      */
     protected $middleware = [
+        'multipartRequestSupport',
         'initializeProcessor',
         'pluginsProcessor',
         'themesProcessor',
@@ -259,6 +261,9 @@ class Grav extends Container
 
         $container = new Container(
             [
+                'multipartRequestSupport' => function () {
+                    return new MultipartRequestSupport();
+                },
                 'initializeProcessor' => function () {
                     return new InitializeProcessor($this);
                 },