+++ system/src/Grav/Common/Media/Interfaces/MediaObjectInterface.php
@@ -159,14 +159,6 @@ interface MediaObjectInterface extends \Grav\Framework\Media\Interfaces\MediaObj
     public function thumbnail($type = 'auto');
 
     /**
-     * Return URL to file.
-     *
-     * @param bool $reset
-     * @return string
-     */
-    public function url($reset = true);
-
-    /**
      * Turn the current Medium into a Link
      *
      * @param  bool $reset
@@ -220,18 +212,6 @@ interface MediaObjectInterface extends \Grav\Framework\Media\Interfaces\MediaObj
      */
     #[\ReturnTypeWillChange]
     public function __call($method, $args);
-
-    /**
-     * Get value by using dot notation for nested arrays/objects.
-     *
-     * @example $value = $this->get('this.is.my.nested.variable');
-     *
-     * @param string $name Dot separated path to the requested value.
-     * @param mixed $default Default value (or null).
-     * @param string|null $separator Separator, defaults to '.'
-     * @return mixed Value.
-     */
-    public function get($name, $default = null, $separator = null);
 
     /**
      * Set value by using dot notation for nested arrays/objects.