+++ system/src/Grav/Framework/RequestHandler/Middlewares/Exceptions.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 
 /**
  * @package    Grav\Framework\RequestHandler
@@ -6,8 +6,6 @@
  * @copyright  Copyright (c) 2015 - 2022 Trilby Media, LLC. All rights reserved.
  * @license    MIT License; see LICENSE file for details.
  */
-
-declare(strict_types=1);
 
 namespace Grav\Framework\RequestHandler\Middlewares;